Refactor image management into custom hook
